+++ examples/spot/Account/Withdraw.php
@@ -1,20 +1,15 @@
 <?php
-use BitMart\Lib\CloudConfig;
 use BitMart\Spot\APIAccount;
-use BitMart\Spot\APISpot;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../ExampleConfig.php';
 
-$APIAccount = new APIAccount(new CloudConfig([
-    'accessKey' => "<your_api_key>",
-    'secretKey' => "<your_secret_key>",
-    'memo' => "<your_memo>",
-]));
+$APIAccount = new APIAccount(ExampleConfig::getExampleConfig());
 
 // Parameters for Withdraw to the blockchain
 $response = $APIAccount->postWithdraw(
-    'USDT-ERC20',
-    '40',
+    'USDT-ETH',
+    '4000',
     [
         'address' => "0xe57b69a8776b37860407965B73cdFFBDF",
         'address_memo' => '123',