+++ examples/spot/Trading/NewBatchOrder.php
@@ -1,14 +1,10 @@
 <?php
-use BitMart\Lib\CloudConfig;
 use BitMart\Spot\APISpot;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../ExampleConfig.php';
 
-$APISpot = new APISpot(new CloudConfig([
-    'accessKey' => "<your_api_key>",
-    'secretKey' => "<your_secret_key>",
-    'memo' => "<your_memo>",
-]));
+$APISpot = new APISpot(ExampleConfig::getExampleConfig());
 
 $orderParam = array();
 
@@ -17,9 +13,10 @@ $orderParam[] =
         'symbol' => 'BTC_USDT',
         'side' => 'buy',
         'type' => 'limit',
-        'size' => '0.1',
+        'size' => '1000000000',
         'price' => '8800',
-        'clientOrderId' => 'test0000000003'
+        'clientOrderId' => 'test0000000003',
+        'stpMode' => "cancel_maker"
     ];
 
 $orderParam[] =
@@ -27,13 +24,14 @@ $orderParam[] =
         'symbol' => 'BTC_USDT',
         'side' => 'sell',
         'type' => 'market',
-        'size' => '0.2',
+        'size' => '2000000000',
         'notional' => '5',
         'clientOrderId' => 'test0000000004'
     ];
 
+// Submit batch order with stpMode
 $response = $APISpot->postSubmitBatchOrder('BTC_USDT', $orderParam, [
-    'recvWindow' => 5000
+    'recvWindow' => 5000,
 ])['response'];
 
 echo json_encode($response);