+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Changelog
   - `/contract/private/position` Get Current Position (KEYED): Add the Request field **account**
   - `/contract/private/order-history` Get Order History (KEYED): Add the Request field **account**, **order_id**, **client_order_id**
   - `/contract/private/order` Get Order Detail (KEYED): Add the Request field **account**
+  - `/account/v1/wallet` Get Account Balance (KEYED): Add the Request Parameters **needUsdValuation** 
+  - `/spot/v4/batch_orders` New Batch Order(v4) (SIGNE): Add the Request Parameters **stpMode** 
+  - `/spot/v2/submit_order New Order(v2) (SIGNED)`: Add the Request Parameters **stpMode** 
+  
 #### Improvements
 - **Enhanced Logging System**: Introduced a comprehensive logging mechanism to replace the previous `echo` and `print_r` output methods
   - Added `CloudLogger` class with configurable logging capabilities
@@ -50,6 +54,7 @@ Changelog
   - `/contract/private/modify-plan-order` Modify Plan Order (SIGNED): Remove the Request Parameters **client_order_id**
   - `/account/v1/currencies` Get Currencies: Add the Request Parameters **startTime** And **endTime**
   - `/account/v2/deposit-withdraw/history` Get Currencies: Add the Request Parameters **currencies**
+
 
 ---